ci: Add Release Workflow
Add a GitHub Workflow for the release scenario. Currently works by
running after a push to main (either by a direct commit or pull request)
that affects one or more AssemblyInfo.cs files.

This workflow will run the added Python script to validate that all of
the versions are in sync, as well as ensure there is no conflicting tag.

If both prerequisites are met, it will extract the version from
AssemblyInfo and create both a tag and a release, uploading the .7z and
.zip archives to the release. The release is currently set to be created
as a pre-release and will need to be manually published, however that
can be changed in the future.

Closes #1