Change updater to use zip file, only release zip file
Currently, ACT get plugins uses the zip file (with 1 layer of folders)
and the OverlayPlugin updater uses the 7z file (with 0 extra layers).
ACT update is also very static, in that it always looks for the second
asset, which must be the zip file.  This is all a bit fragile.

To make this a little bit easier, change the updater to use the zip
file instead, and only release the zip file.  This requires stripping
an extra dir.  Verified by changing the version back in time and
using ngld/OverlayPlugin to update from the zip file properly.