Add generic memory version sorter
Rather than adding explicit logic to every single FooMemoryManager
for which FooMemoryXX to use, make FooMemoryXX implement a new
IVersionedMemory interface which can be sorted generically.

This will let the best candidate be picked first (for performance)
and then after that checking other candidates so that OverlayPlugin
will continue to work for updated regions in the short term
without an OverlayPlugin update.

For performance reasons, we should still update OverlayPlugin's
built-in version when we can, but this will help with transitions.