Load and cache opcodes remotely when they're missing
If code attempts to access an opcode and either it or the game
version is missing, OP will now fetch the opcodes from github.
It will only attempt to do this fetch once per ACT load,
even if the fetch fails for any reason.

It will continue to prefer using this fetched version from github
until either OP's version updates or an opcode is missed again.

Since this is sort of "network activity on behalf of a user who hasn't
requested it", only do this work if the user has opted into automatic
updates (which does the same sort of work).

Closes #65.